+++ integration-tests/forklift/src/storage.spec.ts
@@ -1,7 +1,76 @@
 import { describe, expect, it } from "vitest"
 import { ALICE, getForkliftProvider } from "./lib/forklift"
-import { createClient } from "polkadot-api"
+import { randomBytes } from "crypto"
+import { AccountId, createClient, HexString } from "polkadot-api"
 import { dotAh } from "@polkadot-api/descriptors"
+
+import { methods, ForkliftOptions } from "@polkadot-api/forklift"
+import { JsonRpcRequest } from "@polkadot-api/json-rpc-provider"
+
+type RpcMethod = ForkliftOptions["rpcOverrides"][string]
+
+const chainHead_v1_storage: RpcMethod = async (
+  con,
+  req: JsonRpcRequest<
+    [
+      followSubscription: string,
+      hash: HexString,
+      items: Array<{
+        key: HexString
+        type:
+          | "value"
+          | "hash"
+          | "closestDescendantMerkleValue"
+          | "descendantsValues"
+          | "descendantsHashes"
+      }>,
+      childTrie: HexString | null,
+    ]
+  >,
+  ctx,
+) => {
+  const [followSubscription, hash, items, childTrie] = req.params!
+
+  await new Promise((res) => setTimeout(res, 0))
+  const sub = con.context.chainHead_v1_subs[followSubscription]
+
+  if (!sub || Object.keys(sub.operations).length > 0) {
+    return con.send({
+      jsonrpc: "2.0",
+      id: req.id!,
+      result: { result: "limitReached" },
+    })
+  }
+
+  if (items.length > 32) {
+    const newItems = items.slice(0, 32)
+    return methods.chainHead_v1_storage(
+      {
+        ...con,
+        send(msg) {
+          if ("result" in msg && msg.result.result === "started") {
+            con.send({
+              ...msg,
+              result: {
+                ...msg.result,
+                discardedItems: items.length - 32,
+              },
+            })
+          } else {
+            con.send(msg)
+          }
+        },
+      },
+      {
+        ...req,
+        params: [followSubscription, hash, newItems, childTrie],
+      },
+      ctx,
+    )
+  }
+
+  return methods.chainHead_v1_storage(con, req, ctx)
+}
 
 describe("storage", () => {
   it("subscribes and decodes storage values", async () => {
@@ -21,6 +90,29 @@ describe("storage", () => {
     const entries = await api.query.Referenda.ReferendumInfoFor.getEntries()
     expect(entries.length).toBeGreaterThan(0)
     entries.forEach(({ keyArgs }) => expect(keyArgs[0]).toBeTypeOf("number"))
+
+    client.destroy()
+  })
+
+  it("handles storage query limits", async () => {
+    const [provider, forklift] = getForkliftProvider("storage_ent")
+    forklift.changeOptions({
+      rpcOverrides: {
+        chainHead_v1_storage,
+      },
+    })
+    const client = createClient(provider)
+    const api = client.getTypedApi(dotAh)
+
+    const addresses = Array(70)
+      .fill(null)
+      .map(() => AccountId().dec(randomBytes(32)))
+
+    console.log(`querying ${addresses.length} identities`)
+    const result = await api.query.System.Account.getValues(
+      addresses.map((address) => [address]),
+    )
+    expect(result.length).toEqual(addresses.length)
 
     client.destroy()
   })